renamed trans.{cpp,hpp{ to CodeGenerator.{cpp,hpp} respectively
