+++ src/Makefile.am
@@ -4,7 +4,7 @@ LEXERSOURCES = njnr.tab.cpp lex.yy.cc location.hh njnr.tab.hpp
 njnrSOURCES = command.cpp compiler_core.cpp cpptypes.cpp debug.cpp
 njnrSOURCES += getopt.cpp helper.cpp lex.cpp list.cpp
 njnrSOURCES += main.cpp operator.cpp parser_functions.cpp
-njnrSOURCES += trans.cpp common/symbol_table.cpp common/symbol_table_stack.cpp
+njnrSOURCES += CodeGenerator.cpp common/symbol_table.cpp common/symbol_table_stack.cpp
 njnrSOURCES += common/symbol_table_stackX.cpp common/symbol_table_entry.cpp
 njnrSOURCES += expressions/expressions.cpp
 njnrSOURCES += functions/function_calls.cpp functions/functions.cpp
@@ -12,7 +12,7 @@ njnrSOURCES += statements/statements.cpp
 njnrSOURCES += include/debug.hpp include/type.hpp include/list.hpp include/cpptypes.hpp
 njnrSOURCES += include/constant.hpp include/common/symbol_table_stackX.hpp
 njnrSOURCES += include/common/symbol_table_stack.hpp include/common/symbol_table.hpp
-njnrSOURCES += include/compiler.hpp include/trans.hpp include/common/symbol_table_entry.hpp
+njnrSOURCES += include/compiler.hpp include/CodeGenerator.hpp include/common/symbol_table_entry.hpp
 njnrSOURCES += include/lex.hpp include/data.hpp include/helper.hpp
 njnrSOURCES += include/operator.hpp include/common/symbol_table_templates.h
 njnrSOURCES += include/common/symbol_table_stack_templates.h
